backend/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
dnspython==2.7.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
